refactored the reply flow to remove duplicated effort
